changes the file name
